GigeVisionLibrary.Test.Avalonia Renamed
