+++ GigeVisionLibrary.Avalonia/GigeVisionLibrary.Test.Avalonia.csproj
